general/pdf-table-extractor/requirements.txt
camelot-py[cv]
tabula-py